// packages/core/src/compiler/parser/ast.d.ts
export interface AstNodeBase {
    readonly kind: AstNodeKind;
    readonly location: SourceLocation;
}
export declare enum AstNodeKind {
    Program = "Program",
    FunctionDeclaration = "FunctionDeclaration",
    Parameter = "Parameter",
    Block = "Block",
    VariableDeclaration = "VariableDeclaration",
    ReturnStatement = "ReturnStatement",
    IfStatement = "IfStatement",
    ExpressionStatement = "ExpressionStatement",
    AssignmentExpression = "AssignmentExpression",
    BinaryExpression = "BinaryExpression",
    UnaryExpression = "UnaryExpression",
    Identifier = "Identifier",
    NumberLiteral = "NumberLiteral",
    HexLiteral = "HexLiteral",
    BinaryLiteral = "BinaryLiteral",
    BooleanLiteral = "BooleanLiteral",
    ArrayLiteral = "ArrayLiteral",
    ArrayAccess = "ArrayAccess",
    SliceAccess = "SliceAccess",
    TypeAnnotation = "TypeAnnotation"
}
// ...
export type BinaryOperator = '+' | '-' | '*' | '&' | '|' | '^' | '>>' | '>>>' | '<<' | '===' | '!==' | '>' | '<' | '>=' | '<=';
export type UnaryOperator = '~' | '-' | '!';
export interface AssignmentExpressionNode extends AstNodeBase {
    kind: AstNodeKind.AssignmentExpression;
    target: IdentifierNode | ArrayAccessNode;
    value: ExpressionNode;
}
export interface BinaryExpressionNode extends AstNodeBase {
    kind: AstNodeKind.BinaryExpression;
    operator: BinaryOperator;
    left: ExpressionNode;
    right: ExpressionNode;
}
export interface UnaryExpressionNode extends AstNodeBase {
    kind: AstNodeKind.UnaryExpression;
    operator: UnaryOperator;
    operand: ExpressionNode;
}
export interface IdentifierNode extends AstNodeBase {
    kind: AstNodeKind.Identifier;
    name: string;
}
export interface NumberLiteralNode extends AstNodeBase {
    kind: AstNodeKind.NumberLiteral;
    value: number;
    raw: string;
}
export interface HexLiteralNode extends AstNodeBase {
    kind: AstNodeKind.HexLiteral;
    value: number;
    raw: string;
}
export interface BinaryLiteralNode extends AstNodeBase {
    kind: AstNodeKind.BinaryLiteral;
    value: number;
    raw: string;
}
export interface BooleanLiteralNode extends AstNodeBase {
    kind: AstNodeKind.BooleanLiteral;
    value: boolean;
}
export interface ArrayLiteralNode extends AstNodeBase {
    kind: AstNodeKind.ArrayLiteral;
    elements: ExpressionNode[];
}
export interface ArrayAccessNode extends AstNodeBase {
    kind: AstNodeKind.ArrayAccess;
    array: IdentifierNode;
    index: ExpressionNode;
}
export interface SliceAccessNode extends AstNodeBase {
    kind: AstNodeKind.SliceAccess;
    source: ExpressionNode;
    msb: ExpressionNode;
    lsb: ExpressionNode;
}
export type ExpressionNode = AssignmentExpressionNode | BinaryExpressionNode | UnaryExpressionNode | IdentifierNode | NumberLiteralNode | HexLiteralNode | BinaryLiteralNode | BooleanLiteralNode | ArrayLiteralNode | ArrayAccessNode | SliceAccessNode;